+++ HISTORY.md
@@ -2,8 +2,9 @@ History
 ===============
 
 ### By now 2014-03-16
-* 1 minor enhancement
+* 2 minor enhancement
 	* Improved BioTCM::Table row insertion
+	* Improved BioTCM::Databases::Cipher performance
 
 ### 0.0.2 2014-03-16
 * 3 major enhancement